Moved suggester tests into own test file from ConceptGraphTest
