+++ test/edu/ithaca/dragonlab/ckc/conceptgraph/ConceptGraphTest.java
@@ -1,21 +1,16 @@
 package edu.ithaca.dragonlab.ckc.conceptgraph;
 
-
 import edu.ithaca.dragonlab.ckc.io.ConceptGraphRecord;
 import edu.ithaca.dragonlab.ckc.io.LearningObjectLinkRecord;
 import edu.ithaca.dragonlab.ckc.learningobject.ExampleLearningObjectFactory;
 import edu.ithaca.dragonlab.ckc.learningobject.ExampleLearningObjectResponseFactory;
 import edu.ithaca.dragonlab.ckc.learningobject.LearningObject;
-
 import edu.ithaca.dragonlab.ckc.learningobject.LearningObjectResponse;
-import edu.ithaca.dragonlab.ckc.suggester.LearningObjectSuggester;
-import edu.ithaca.dragonlab.ckc.suggester.LearningObjectSuggestion;
 import edu.ithaca.dragonlab.ckc.util.DataUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.Assert;
 import org.junit.Test;
-
 
 import java.util.*;
 import java.util.ArrayList;
@@ -23,41 +18,8 @@ import java.util.List;
 
 import static org.hamcrest.collection.IsIterableContainingInAnyOrder.containsInAnyOrder;
 
-
 public class ConceptGraphTest {
 	static Logger logger = LogManager.getLogger(ConceptGraphTest.class);
-
-
-    @Test
-    public void SuggestedConceptNodeMapSimpleTest() {
-        ConceptGraph orig = ExampleConceptGraphFactory.simpleTestGraphTest();
-
-        //what it is
-        HashMap<String, List<LearningObjectSuggestion>> objectSuggestionMap = LearningObjectSuggester.SuggestedConceptNodeMap(orig);
-
-        //what it should be
-        List<LearningObjectSuggestion> testList3 = new ArrayList<>();
-        Assert.assertEquals(1, objectSuggestionMap.size());
-        Assert.assertEquals(testList3, objectSuggestionMap.get("C"));
-
-
-
-    }
-
-
-    @Test
-    public void SuggestedConceptNodeMapWillOneStudentTest() {
-        ConceptGraph orig= ExampleConceptGraphFactory.willExampleConceptGraphTestOneStudent();
-        //what it is
-        HashMap<String, List<LearningObjectSuggestion>> objectSuggestionMap = LearningObjectSuggester.SuggestedConceptNodeMap(orig);
-
-        Assert.assertEquals(2, objectSuggestionMap.size());
-        Assert.assertEquals(objectSuggestionMap.get("If Statement").get(1).getId(), "Q3");
-        Assert.assertEquals(objectSuggestionMap.get("If Statement").get(0).getId(), "Q10");
-        Assert.assertEquals(objectSuggestionMap.get("While Loop").get(0).getId(), "Q10");
-
-    }
-
 
     @Test
     public void buildLearningObjectListSimpleTest(){
@@ -97,60 +59,6 @@ public class ConceptGraphTest {
 
         //from a node not in graph
         Assert.assertEquals(null, orig.buildLearningObjectSummaryList("W"));
-    }
-
-
-
-
-    @Test
-    public void suggestedOrderBuildLearningObjectListTest(){
-        List<LearningObjectLinkRecord> myList = ExampleLearningObjectFactory.makeSimpleLOLRecords();
-        myList.add(new LearningObjectLinkRecord("Q10", Arrays.asList("A"),1));
-        ConceptGraph orig = new ConceptGraph(ExampleConceptGraphRecordFactory.makeSimple(),
-                myList, ExampleLearningObjectResponseFactory.makeSimpleResponses());
-
-//        //from A nodes
-        HashMap<String, Integer> testCompareA = new HashMap<String, Integer>();
-        testCompareA.put("Q3",2);
-        testCompareA.put("Q4",2);
-        testCompareA.put("Q5",2);
-        testCompareA.put("Q6",2);
-        testCompareA.put("Q1",1);
-        testCompareA.put("Q2",1);
-        testCompareA.put("Q10",1);
-
-        HashMap<String, Integer> learningSummaryFromA = orig.buildLearningObjectSummaryList("A");
-        //makes sure that buildLearningObjectSummaryList works
-        Assert.assertEquals(testCompareA,learningSummaryFromA);
-
-        //build the sugggested learning object list
-        List<LearningObjectSuggestion> suggestedList = LearningObjectSuggester.buildLearningObjectSuggestionList(learningSummaryFromA, orig.getLearningObjectMap());
-
-
-        //this is ordered based on "level"
-        List<LearningObjectSuggestion> suggestListTest = new ArrayList<>();
-        suggestListTest.add(new LearningObjectSuggestion("Q3", 2, LearningObjectSuggestion.Level.RIGHT) );
-        suggestListTest.add(new LearningObjectSuggestion("Q1", 1, LearningObjectSuggestion.Level.RIGHT) );
-        suggestListTest.add(new LearningObjectSuggestion("Q2", 1, LearningObjectSuggestion.Level.RIGHT) );
-        suggestListTest.add(new LearningObjectSuggestion("Q4", 2, LearningObjectSuggestion.Level.WRONG) );
-        suggestListTest.add(new LearningObjectSuggestion("Q5", 2, LearningObjectSuggestion.Level.WRONG) );
-        suggestListTest.add(new LearningObjectSuggestion("Q6", 2, LearningObjectSuggestion.Level.WRONG) );
-        suggestListTest.add(new LearningObjectSuggestion("Q10", 1, LearningObjectSuggestion.Level.INCOMPLETE) );
-
-
-        //orders the list based off of "level"
-        LearningObjectSuggester.suggestedOrderBuildLearningObjectList(suggestedList);
-
-//        //who should call orderedList
-
-        for (int i =0; i<suggestedList.size(); i++){
-
-            Assert.assertEquals(suggestedList.get(i).getId(),suggestListTest.get(i).getId());
-            Assert.assertEquals(suggestedList.get(i).getPathNum(),suggestListTest.get(i).getPathNum());
-            Assert.assertEquals(suggestedList.get(i).getLevel(),suggestListTest.get(i).getLevel());
-
-        }
-
     }
 
 
@@ -204,8 +112,6 @@ public class ConceptGraphTest {
         Assert.assertEquals(2, graph.getLearningObjectMap().get("Q6").getResponses().size());
     }
 
-    //TODO: Write a test where a single learning object is connected to multiple ConceptNodes, makes sure its the same learning object by address
-
     @Test
     public void addLearningObjectsTest(){
         ConceptGraph graph = ExampleConceptGraphFactory.makeSimpleWithData();
@@ -233,70 +139,6 @@ public class ConceptGraphTest {
         Assert.assertEquals(0.5,   graph.findNodeById("C").getLearningObjectMap().get("Q6").calcKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
     }
 
-
-    //TODO
-	@Test
-	public void makeConceptGraphFromFileTest(){
-//		ObjectMapper mapper = new ObjectMapper();
-//		mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-//
-//        try {
-//			ConceptGraphRecord lists = mapper.readValue(new File("test/testresources/ABCSimple.json"), ConceptGraphRecordOld.class);
-//			Assert.assertEquals(11, lists.getNodes().size());
-//			Assert.assertEquals(11, lists.getLinks().size());
-//
-//			ConceptGraph myGraph = new ConceptGraph(lists);
-//			ConceptGraph myTree = myGraph.graphToTree();
-//
-//			ConceptGraphRecordOld listsFromTree = myTree.buildConceptGraphRecord();
-//
-//			Assert.assertEquals(16, listsFromTree.getNodes().size());
-//			Assert.assertEquals(15, listsFromTree.getLinks().size());
-//
-//		} catch (IOException e) {
-//			e.printStackTrace();
-//		}
-		
-	}
-
-    //TODO
-    @Test
-	public void makeJsonFromConceptGraphTreeTest(){
-
-//		ObjectMapper mapper = new ObjectMapper();
-//		mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-//
-//		String outputLocation = "out/test/sampleoutput/CarrieJsonGraph.json";
-//
-//        try {
-//			ConceptGraphRecordOld lists = mapper.readValue(new File("test/testresources/ABCSimple.json"), ConceptGraphRecordOld.class);
-//			ConceptGraph tree = new ConceptGraph(lists).graphToTree();
-//
-//			try{
-//			mapper.writeValue(new File(outputLocation), tree.buildConceptGraphRecord());
-//            } catch (Exception e) {
-//                e.printStackTrace();
-//            }
-//        } catch (IOException e) {
-//			e.printStackTrace();
-//		}
-//
-//		String jsonString = "";
-//		//reads json from file
-//		try {
-//			for (String line : Files.readAllLines(Paths.get(outputLocation))) {
-//			    jsonString += line;
-//			}
-//		} catch (IOException e) {
-//			e.printStackTrace();
-//		}
-//
-//		int numID = jsonString.split("id").length-1;
-//		int numLink = jsonString.split("parent").length-1;
-//
-//		Assert.assertEquals(16, numID);
-//		Assert.assertEquals(15, numLink);
-	}
     @Test
     public void linkLearningObjectsTest(){
         ConceptGraph graph = ExampleConceptGraphFactory.makeSimpleWithData();
@@ -338,7 +180,7 @@ public class ConceptGraphTest {
     }
 
     @Test
-    public void addLearningObjectsFromLearningObjectLinkRecords(){
+    public void addLearningObjectsFromLearningObjectLinkRecordsTest(){
         //Creating graph
         ConceptGraph graph = ExampleConceptGraphFactory.makeSimpleWithData();
 
@@ -369,8 +211,17 @@ public class ConceptGraphTest {
         Assert.assertEquals("Q7",graph.getLearningObjectMap().get("Q7").getId());
         // Makes sure the new question was only added once to Learning Object map (previously had 6 questions, now 7)
         Assert.assertEquals(7,graph.getLearningObjectMap().size());
+    }
 
+	@Test
+	public void calcKnowledgeEstimateTest(){
+        ConceptGraph graph = ExampleConceptGraphFactory.makeSimpleWithData();
+        graph.calcDataImportance();
+        graph.calcKnowledgeEstimates();
 
+        Assert.assertEquals(0.5, graph.findNodeById("C").getKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
+        Assert.assertEquals(0.688, graph.findNodeById("B").getKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
+        Assert.assertEquals(0.615, graph.findNodeById("A").getKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
     }
 
     @Test
@@ -386,61 +237,6 @@ public class ConceptGraphTest {
         Assert.assertEquals(0.7666, orig.findNodeById("For Loop").getKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
 
     }
-
-
-	@Test
-	public void calcKnowledgeEstimateTest(){
-        ConceptGraph graph = ExampleConceptGraphFactory.makeSimpleWithData();
-        graph.calcDataImportance();
-        graph.calcKnowledgeEstimates();
-
-        Assert.assertEquals(0.5, graph.findNodeById("C").getKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
-        Assert.assertEquals(0.688, graph.findNodeById("B").getKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
-        Assert.assertEquals(0.615, graph.findNodeById("A").getKnowledgeEstimate(), DataUtil.OK_FLOAT_MARGIN);
-    }
-
-    //TODO: Adapt to new graph creation once data is reinstated
-//	@Test
-//	public void calcActualCompTest() {
-//        //create your selectionGraph and add responses to it
-//
-//		float delta = (float) .01;
-//
-//		//Checks children on If Statement to make sure their actualComps are correct
-//		//(gotten from visualized graph)
-//		//file:///Users/willsuchanek/git/metafora-project/MonitorInterventionMetafora/war/TreeDisplay/index.html
-//		Assert.assertEquals(0, selectionListMap.get("Boolean").getKnowledgeEstimate(),delta);
-//		Assert.assertEquals(1, selectionListMap.get("test_question6_4_2").getKnowledgeEstimate(),delta);
-//		Assert.assertEquals(1, selectionListMap.get("test_question6_4_1").getKnowledgeEstimate(),delta);
-//
-//
-//		//Calculates the actualComp for If statements children
-//		float sumOfChildren = 0;
-//		int childCounter = 0;
-//		for(ConceptNode node :selectionListMap.get("If Statement").getChildren()){
-//			sumOfChildren+=node.getKnowledgeEstimate();
-//			childCounter++;
-//		}
-//		//makes sure the calculated number is equal to the number stored in If Statements
-//		Assert.assertEquals(sumOfChildren/childCounter, selectionListMap.get("If Statement").getKnowledgeEstimate(),delta);
-//
-//		//Checks children on Control to make sure their actualComps are correct
-//		//(gotten from visualized graph)
-//		//file:///Users/willsuchanek/git/metafora-project/MonitorInterventionMetafora/war/TreeDisplay/index.html
-//		Assert.assertEquals(.67, selectionListMap.get("If Statement").getKnowledgeEstimate(),delta);
-//		Assert.assertEquals(0, selectionListMap.get("Loops").getKnowledgeEstimate(),delta);
-//
-//		//Calculates the actualComp for Control children
-//		sumOfChildren = 0;
-//		childCounter = 0;
-//		for(ConceptNode node :selectionListMap.get("Control").getChildren()){
-//			sumOfChildren+=node.getKnowledgeEstimate();
-//			childCounter++;
-//		}
-//
-//		//makes sure the calculated number is equal to the number stored in If Statements
-//		Assert.assertEquals(sumOfChildren/childCounter, selectionListMap.get("Control").getKnowledgeEstimate(),delta);
-//	}
 
 }
